+++ lib/datamappify/data/criteria/relational/count.rb
@@ -1,10 +1,12 @@
+require 'datamappify/data/criteria/relational/criteria'
+
 module Datamappify
   module Data
     module Criteria
       module Relational
-        class Count < Common
+        class Count < Criteria
           def perform
-            source_class.count
+            (records || source_class).count
           end
         end
       end