+++ lib/datamappify/repository/query_method/criteria.rb
@@ -1,12 +1,14 @@
+require 'datamappify/repository/collection'
+
 module Datamappify
   module Repository
     module QueryMethod
       class Criteria < Method
         # @return [Array<Entity>]
         def perform
-          results = dispatch_criteria_to_default_source(
+          results = Collection.new(dispatch_criteria_to_default_source(
             :Criteria, data_mapper.entity_class, criteria, data_mapper.attributes
-          )
+          ))
 
           MultiResultBlender.new(self).blend(results)
         end