+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `before_load`, `after_load`, `before_find` and `after_find` callbacks.
 - Added `before_init` and `after_init` callbacks.
 - Added `offset` argument to the `limit` criteria.
+- Added support for Kaminari pagination.
 
 ## 0.70.0.beta1 [2013-09-26]